#131 - refactor: 색테마api명 변경 및 json 데이터 추가
